LISP Programming
This programming assignment emphasizes both the applicative and the recursive aspects of LISP. Carefully implement it using only the LISP primitives COND, CAR, CONS, CDR, NULL, EQUAL, DEFUN, and functions which you define. 

Write and test a LISP function MY-UNION which has two arguments (each a list) and which returns the union of two lists (a list consisting of all elements from both lists, without duplicating elements which occur in both lists). Also write a LISP function MY-INTERSECT which has two arguments (each a list) and which returns the intersection of the two lists (a list consisting of all elements which they both have in common, assuming neither list has duplicate elements). In addition to the MY-MEMBER, MY-UNION, and MY-INTERSECTION functions, graduate students should write and test a function MY-DIFFERENCE which returns the difference of two sets.